+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlobContainerTests.cs
@@ -11,7 +11,7 @@ using Microsoft.WindowsAzure.Storage.Blob;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlobContainerTests : StandaloneAzureTestsBase
     {