+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlobDirectoryListTests.cs
@@ -9,7 +9,7 @@ using Microsoft.WindowsAzure.Storage.Blob;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlobDirectoryListTests : StandaloneAzureTestsBase
     {