+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlobStorageClientTests.cs
@@ -5,7 +5,7 @@ using LightBlue.Standalone;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlobStorageClientTests
     {