+++ LightBlue.Tests/Standalone/BlobStorage/StandaloneAzureBlockBlobDownloadTests.cs
@@ -7,7 +7,7 @@ using LightBlue.Standalone;
 using Xunit;
 using Xunit.Extensions;
 
-namespace LightBlue.Tests.Standalone
+namespace LightBlue.Tests.Standalone.BlobStorage
 {
     public class StandaloneAzureBlockBlobDownloadTests : StandaloneAzureTestsBase
     {