+++ LightBlue/IAzureQueueMessage.cs
@@ -1,9 +1,0 @@
-﻿namespace LightBlue
-{
-    public interface IAzureQueueMessage
-    {
-        byte[] AsBytes { get; }
-        string AsString { get; }
-        int DequeueCount { get; }
-    }
-}