+++ ReleaseNotes.md
@@ -3,6 +3,11 @@ LightBlue
 
 Version 1.1.17
 --------------------
+###Features
+* Queue storage
+
+###Implementation
+* Include PDBs in NuGet packages
 
 Version 1.1.16
 --------------------
@@ -13,9 +18,6 @@ Version 1.1.16
 
 Version 1.1.15
 --------------------
-### Features
-* Support for Azure Queue Storage
-
 ###Implementation
 * Update to NewtonSoft.Json 6.0.8